Fix: Check if the resource is required by quattor before any deletion

+++ ncm-opennebula/src/main/perl/opennebula.pm
@@ -1,4 +1,6 @@
-# #
+# ${license-info}
+# ${developer-info}
+# ${author-info}
 # Software subject to following license(s):
 #   Apache 2 License (http://www.opensource.org/licenses/apache2.0)
 #   Copyright (c) Responsible Organization
@@ -20,18 +22,13 @@ use NCM::Component;
 use base qw(NCM::Component);
 use vars qw(@ISA $EC);
 use LC::Exception;
-# OpenNebula Quattor component requires Net::OpenNebula module
-# available from github: https://github.com/stdweird/p5-net-opennebula
 use Net::OpenNebula;
 
+# TODO use constant from CAF::Render
 use constant TEMPLATEPATH => "/usr/share/templates/quattor";
 
 our $EC=LC::Exception::Context->new->will_store_all;
 
-# Function to connect to ONE RPC endpoint
-# to admin and manage OpenNebula resources 
-# it requires a valid ONE admin user/pass 
-# onadmin is used by default
 sub make_one 
 {
     my ($self, $rpc) = @_;
@@ -53,16 +50,17 @@ sub make_one
     return $one;
 }
 
+# TODO replace by CAF::Render
 # Detect and process ONE templates
 sub process_template 
 {
-    my ($self, $config, $tt_name) = @_;
+    my ($self, $config, $type_name) = @_;
     my $res;
     
-    my $tt_rel = "metaconfig/opennebula/$tt_name.tt";
+    my $type_rel = "metaconfig/opennebula/$type_name.tt";
     my $tpl = Template->new(INCLUDE_PATH => TEMPLATEPATH);
-    if (! $tpl->process($tt_rel, { $tt_name => $config }, \$res)) {
-            $self->error("TT processing of $tt_rel failed:", 
+    if (! $tpl->process($type_rel, { $type_name => $config }, \$res)) {
+            $self->error("TT processing of $type_rel failed:", 
                                           $tpl->error());
             return;
     }
@@ -73,34 +71,34 @@ sub process_template
 # based on resource type
 sub create_something
 {
-    my ($self, $one, $tt, $data) = @_;
+    my ($self, $one, $type, $data) = @_;
     
-    my $template = $self->process_template($data, $tt);
-    my $name;
-    my $new;
+    my $template = $self->process_template($data, $type);
+    my ($name, $new);
     if (!$template) {
-        $self->error("No template data found for $tt.");
+        $self->error("No template data found for $type.");
         return;
     }
     if ($template =~ m/^NAME\s+=\s+(?:"|')(.*?)(?:"|')\s*$/m) {
         $name = $1;
-        $self->verbose("Found template NAME: $name within $tt resource.");
+        $self->verbose("Found template NAME: $name within $type resource.");
     } else {
         $self->error("Template NAME not found.");
         return;
     }
-    my $cmethod = "create_$tt";
+    my $cmethod = "create_$type";
 
-    $self->info("Creating new $name $tt resource.");
-    # Check first if the resource name already exists
-    my @rname = $self->detect_name($one, $tt, $name);
+    my @rname = $self->detect_used_resource($one, $type, $name);
     if (!@rname) {
+        $self->info("Creating new $name $type resource.");
         $new = $one->$cmethod($template);
     }
     return $new;
 }
 
-sub detect_name
+# Detects if the resource
+# is already there
+sub detect_used_resource
 {
     my ($self, $one, $type, $name) = @_;
     my $gmethod = "get_${type}s";
@@ -109,48 +107,65 @@ sub detect_name
 	$self->error("Name: $name is already used by a $type resource. We can't create the same resource twice.");
 	return @existres;
     } else {
-	$self->info("Name: $name is not used by $type resource yet.");
+	$self->verbose("Name: $name is not used by $type resource yet.");
         return;
     }
 }
 
+# Create resource name
+# array list
+sub create_resource_names_list
+{
+    my ($self, $one, $type, $resources) = @_;
+    my ($name,@namelist, $template);
+
+    foreach my $newresource (@$resources) {
+        $template = $self->process_template($newresource, $type);
+        if ($template =~ m/^NAME\s+=\s+(?:"|')(.*?)(?:"|')\s*$/m) {
+            $name = $1;
+            push(@namelist, $name);
+        }
+    }
+    return @namelist;
+}
 
 # Remove/add ONE resources
 # based on resource type
 sub manage_something
 {
     my ($self, $one, $type, $resources) = @_;
-    my $remove;
+    my ($remove, @namelist);
 
     if (!$resources) {
         $self->error("No $type resources found.");
+        return;
     } else {
-        $self->info("Found new $type resources.");
+        $self->verbose("Managing $type resources.");
     }
 
     if (($type eq "kvm") or ($type eq "xen")) {
         $self->manage_hosts($one, $type, $resources);
         return;
-    }
-
-    if ($type eq "user") {
+    } elsif ($type eq "user") {
         $self->manage_users($one, $resources);
         return;
     }
 
-    $self->info("Removing old ${type}/s");
+    $self->verbose("Check to remove ${type}s");
     my $method = "get_${type}s";
     my @existres = $one->$method(qr{^.*$});
+    my @namelist = $self->create_resource_names_list($one, $type, $resources);
+    my %rnames = map { $_ => 1 } @namelist;
     foreach my $oldresource (@existres) {
         # Remove the resource only if the QUATTOR flag is set
         # and they are not being used
         if ($type eq "datastore" and !$oldresource->{extended_data}->{IMAGES}->[0]->{ID}->[0]) {
             $remove = 1;
-        }
+        } 
         if ($type eq "vnet" and !$oldresource->{extended_data}->{TOTAL_LEASES}->[0]) {
             $remove = 1;
         }
-        if ($oldresource->{extended_data}->{TEMPLATE}->[0]->{QUATTOR}->[0] and $remove) {
+        if ($oldresource->{extended_data}->{TEMPLATE}->[0]->{QUATTOR}->[0] and $remove and !exists($rnames{$oldresource->name})) {
             $self->info("Removing old resource: ", $oldresource->name);
             $oldresource->delete();
         } else {
@@ -158,7 +173,9 @@ sub manage_something
         };
     }
 
-    $self->info("Creating new ${type}/s");
+    if (scalar @$resources > 0) {
+        $self->info("Creating new ${type}/s: ", scalar @$resources);
+    }
     foreach my $newresource (@$resources) {
         my $new = $self->create_something($one, $type, $newresource);
     }
@@ -171,10 +188,11 @@ sub manage_hosts
     my $new;
     $self->info("Removing old $type hosts.");
     my @existhost = $one->get_hosts(qr{^.*$});
+    my %newhosts = map { $_ => 1 } @$hosts;
     foreach my $t (@existhost) {
         # Remove the host only if there are no VMs running on it
-        if ($t->{extended_data}->{HOST_SHARE}->[0]->{RUNNING_VMS}->[0]) {
-            $self->error("We can't remove this host. There are still running VMs on host: ", $t->name);
+        if ($t->{extended_data}->{HOST_SHARE}->[0]->{RUNNING_VMS}->[0] or exists($newhosts{$t->name})) {
+            $self->error("We can't remove this $type host. There are still running VMs or is a Quattor host: ", $t->name);
         } else {
             $self->info("Removing $type host: ", $t->name);
             $t->delete();
@@ -188,9 +206,9 @@ sub manage_hosts
             'vmm_mad' => $type, 
             'vnm_mad' => "dummy"
         );
-        my @rname = $self->detect_name($one, "host", $host);
+        my @rname = $self->detect_used_resource($one, "host", $host);
         if (!@rname) {
-            $self->verbose("Creating new $type host $host.");
+            $self->info("Creating new $type host $host.");
             $new = $one->create_host(%host_options);
         }
     }
@@ -201,24 +219,14 @@ sub manage_hosts
 sub manage_users
 {
     my ($self, $one, $users) = @_;
-    my $new;
-
-    my @exitsuser = $one->get_users(qr{^.*$});
-    foreach my $t (@exitsuser) {
-        # Remove the user only if the QUATTOR flag is set
-        if ($t->{extended_data}->{TEMPLATE}->[0]->{QUATTOR}->[0]) {
-            $self->info("Removing old regular user: ", $t->name);
-            $t->delete();
-        } else {
-            $self->error("QUATTOR flag not found. We can't remove this user: ", $t->name);
-        }
-    }
+    my ($new, @userlist);
 
     foreach my $user (@$users) {
         if ($user->{user} && $user->{password}) {
             # TODO: Create users with QUATTOR flag set
             # we have to update the user after its creation
-            my @rname = $self->detect_name($one, "user", $user->{user});
+            my @rname = $self->detect_used_resource($one, "user", $user->{user});
+            push(@userlist, $user->{user});
             if (!@rname) {
                 $self->info("Creating new user: ", $user->{user});
                 $new = $one->create_user($user->{user}, $user->{password}, "core");
@@ -226,6 +234,19 @@ sub manage_users
         }
         else {
             $self->error("No user name or password info available.");
+        }
+    }
+
+    my @exitsuser = $one->get_users(qr{^.*$});
+    my %newusers = map { $_ => 1 } @userlist;
+
+    foreach my $t (@exitsuser) {
+        # Remove the user only if the QUATTOR flag is set
+        if ($t->{extended_data}->{TEMPLATE}->[0]->{QUATTOR}->[0] and !exists($newusers{$t->name})) {
+            $self->info("Removing old regular user: ", $t->name);
+            $t->delete();
+        } else {
+            $self->error("QUATTOR flag not found or this user was already included by Quattor. We can't remove it: ", $t->name);
         }
     }
